Split local sync repository cloud identity tests
