+++ apps/android/data/local/src/androidTest/java/com/flashcardsopensourceapp/data/local/cloudidentity/LocalSyncRepositoryDisconnectedStateTest.kt
@@ -1,0 +1,268 @@
+package com.flashcardsopensourceapp.data.local.cloudidentity
+
+import androidx.test.ext.junit.runners.AndroidJUnit4
+import com.flashcardsopensourceapp.data.local.bootstrap.localWorkspaceName
+import com.flashcardsopensourceapp.data.local.cloud.CloudRemoteException
+import com.flashcardsopensourceapp.data.local.database.CardEntity
+import com.flashcardsopensourceapp.data.local.model.CloudAccountState
+import com.flashcardsopensourceapp.data.local.model.CloudServiceConfigurationMode
+import com.flashcardsopensourceapp.data.local.model.EffortLevel
+import com.flashcardsopensourceapp.data.local.model.FsrsCardState
+import com.flashcardsopensourceapp.data.local.model.SyncStatus
+import kotlinx.coroutines.flow.first
+import kotlinx.coroutines.runBlocking
+import org.json.JSONObject
+import org.junit.After
+import org.junit.Assert.assertEquals
+import org.junit.Assert.assertNull
+import org.junit.Assert.assertTrue
+import org.junit.Before
+import org.junit.Test
+import org.junit.runner.RunWith
+
+@RunWith(AndroidJUnit4::class)
+class LocalSyncRepositoryDisconnectedStateTest {
+    private lateinit var environment: CloudIdentityTestEnvironment
+
+    @Before
+    fun setUp() = runBlocking {
+        environment = CloudIdentityTestEnvironment.create()
+    }
+
+    @After
+    fun tearDown() {
+        environment.close()
+    }
+
+    @Test
+    fun syncDisconnectsCloudIdentityWhenRemoteAccountIsDeletedWithoutResettingLocalState() = runBlocking {
+        val initialLocalWorkspaceId = environment.requireLocalWorkspaceId()
+        val initialInstallationId = environment.cloudPreferencesStore.currentCloudSettings().installationId
+        val remoteGateway = FakeCloudRemoteGateway.forFetchAccountError(
+            fetchAccountError = CloudRemoteException(
+                message = "Cloud request failed with status 410 for /me",
+                statusCode = 410,
+                responseBody = JSONObject()
+                    .put("code", "ACCOUNT_DELETED")
+                    .put("requestId", "request-1")
+                    .toString(),
+                errorCode = "ACCOUNT_DELETED",
+                requestId = "request-1",
+                syncConflict = null
+            )
+        )
+        val repository = environment.createSyncRepository(remoteGateway = remoteGateway)
+
+        environment.prepareLinkedCloudIdentity(localWorkspaceId = initialLocalWorkspaceId)
+        environment.aiChatPreferencesStore.updateConsent(hasConsent = true)
+
+        repository.syncNow()
+
+        val localWorkspace = requireNotNull(environment.database.workspaceDao().loadAnyWorkspace()) {
+            "Expected a local workspace after remote deletion."
+        }
+
+        assertEquals(CloudAccountState.DISCONNECTED, environment.cloudPreferencesStore.currentCloudSettings().cloudState)
+        assertEquals(initialInstallationId, environment.cloudPreferencesStore.currentCloudSettings().installationId)
+        assertEquals(
+            com.flashcardsopensourceapp.data.local.model.AccountDeletionState.Hidden,
+            environment.cloudPreferencesStore.currentAccountDeletionState()
+        )
+        assertNull(environment.cloudPreferencesStore.loadCredentials())
+        assertTrue(environment.aiChatPreferencesStore.hasConsent())
+        assertEquals(SyncStatus.Idle, repository.observeSyncStatus().first().status)
+        assertEquals(localWorkspaceName, localWorkspace.name)
+        assertEquals(initialLocalWorkspaceId, localWorkspace.workspaceId)
+    }
+
+    @Test
+    fun syncKeepsDisconnectedStateWhenStoredGuestSessionExistsButCloudStateIsDisconnected() = runBlocking {
+        val localWorkspaceId = environment.requireLocalWorkspaceId()
+        val guestWorkspaceId = "guest-workspace"
+        val remoteGateway = FakeCloudRemoteGateway.standard()
+        val syncRepository = environment.createSyncRepository(remoteGateway = remoteGateway)
+        environment.cloudPreferencesStore.updateCloudSettings(
+            cloudState = CloudAccountState.DISCONNECTED,
+            linkedUserId = null,
+            linkedWorkspaceId = null,
+            linkedEmail = null,
+            activeWorkspaceId = localWorkspaceId
+        )
+        environment.guestAiSessionStore.saveSession(
+            localWorkspaceId = guestWorkspaceId,
+            session = createStoredGuestAiSession(
+                workspaceId = guestWorkspaceId,
+                configurationMode = CloudServiceConfigurationMode.OFFICIAL,
+                apiBaseUrl = "https://api.flashcards-open-source-app.com/v1",
+                guestToken = "guest-token",
+                userId = "guest-user"
+            )
+        )
+
+        try {
+            syncRepository.syncNow()
+        } catch (_: IllegalStateException) {
+        }
+
+        val cloudSettings = environment.cloudPreferencesStore.currentCloudSettings()
+        assertEquals(CloudAccountState.DISCONNECTED, cloudSettings.cloudState)
+        assertEquals(localWorkspaceId, cloudSettings.activeWorkspaceId)
+        assertEquals(localWorkspaceId, environment.database.workspaceDao().loadAnyWorkspace()?.workspaceId)
+        assertTrue(syncRepository.observeSyncStatus().first().status is SyncStatus.Failed)
+        assertTrue(remoteGateway.bootstrapPullWorkspaceIds.isEmpty())
+    }
+
+    @Test
+    fun syncKeepsDisconnectedStateWhenStoredCredentialsExistButCloudStateIsDisconnected() = runBlocking {
+        val localWorkspaceId = environment.requireLocalWorkspaceId()
+        val remoteGateway = FakeCloudRemoteGateway.forAccountSnapshot(
+            accountSnapshot = createCloudAccountSnapshot(
+                userId = "user-1",
+                email = "user@example.com",
+                workspaces = listOf(
+                    createCloudWorkspaceSummary(
+                        workspaceId = "workspace-linked",
+                        name = "Personal",
+                        createdAtMillis = 200L,
+                        isSelected = true
+                    )
+                )
+            )
+        )
+        val syncRepository = environment.createSyncRepository(remoteGateway = remoteGateway)
+        environment.cloudPreferencesStore.saveCredentials(
+            createStoredCloudCredentials(idTokenExpiresAtMillis = Long.MAX_VALUE)
+        )
+        environment.cloudPreferencesStore.updateCloudSettings(
+            cloudState = CloudAccountState.DISCONNECTED,
+            linkedUserId = null,
+            linkedWorkspaceId = null,
+            linkedEmail = null,
+            activeWorkspaceId = localWorkspaceId
+        )
+
+        try {
+            syncRepository.syncNow()
+        } catch (_: IllegalStateException) {
+        }
+
+        val cloudSettings = environment.cloudPreferencesStore.currentCloudSettings()
+        assertEquals(CloudAccountState.DISCONNECTED, cloudSettings.cloudState)
+        assertEquals(localWorkspaceId, cloudSettings.activeWorkspaceId)
+        assertNull(cloudSettings.linkedWorkspaceId)
+        assertNull(cloudSettings.linkedEmail)
+        assertEquals(localWorkspaceId, environment.database.workspaceDao().loadAnyWorkspace()?.workspaceId)
+        assertTrue(syncRepository.observeSyncStatus().first().status is SyncStatus.Failed)
+        assertTrue(remoteGateway.bootstrapPullWorkspaceIds.isEmpty())
+    }
+
+    @Test
+    fun syncKeepsDisconnectedStateWhenLocalShellContainsData() = runBlocking {
+        val localWorkspaceId = environment.requireLocalWorkspaceId()
+        val remoteGateway = FakeCloudRemoteGateway.standard()
+        val syncRepository = environment.createSyncRepository(remoteGateway = remoteGateway)
+        environment.cloudPreferencesStore.saveCredentials(
+            createStoredCloudCredentials(idTokenExpiresAtMillis = Long.MAX_VALUE)
+        )
+        environment.cloudPreferencesStore.updateCloudSettings(
+            cloudState = CloudAccountState.DISCONNECTED,
+            linkedUserId = null,
+            linkedWorkspaceId = null,
+            linkedEmail = null,
+            activeWorkspaceId = localWorkspaceId
+        )
+        environment.database.cardDao().insertCard(
+            CardEntity(
+                cardId = "card-1",
+                workspaceId = localWorkspaceId,
+                frontText = "Question",
+                backText = "Answer",
+                effortLevel = EffortLevel.FAST,
+                dueAtMillis = null,
+                createdAtMillis = 100L,
+                updatedAtMillis = 100L,
+                reps = 0,
+                lapses = 0,
+                fsrsCardState = FsrsCardState.NEW,
+                fsrsStepIndex = null,
+                fsrsStability = null,
+                fsrsDifficulty = null,
+                fsrsLastReviewedAtMillis = null,
+                fsrsScheduledDays = null,
+                deletedAtMillis = null
+            )
+        )
+
+        try {
+            syncRepository.syncNow()
+        } catch (_: IllegalStateException) {
+        }
+
+        val cloudSettings = environment.cloudPreferencesStore.currentCloudSettings()
+        assertEquals(CloudAccountState.DISCONNECTED, cloudSettings.cloudState)
+        assertEquals(localWorkspaceId, cloudSettings.activeWorkspaceId)
+        assertEquals(localWorkspaceId, environment.database.workspaceDao().loadAnyWorkspace()?.workspaceId)
+        assertTrue(remoteGateway.bootstrapPullWorkspaceIds.isEmpty())
+    }
+
+    @Test
+    fun syncDisconnectsInvalidGuestStateWhenStoredGuestSessionIsMissing() = runBlocking {
+        val localWorkspaceId = environment.requireLocalWorkspaceId()
+        val initialInstallationId = environment.cloudPreferencesStore.currentCloudSettings().installationId
+        val syncRepository = environment.createSyncRepository(
+            remoteGateway = FakeCloudRemoteGateway.standard()
+        )
+        environment.cloudPreferencesStore.updateCloudSettings(
+            cloudState = CloudAccountState.GUEST,
+            linkedUserId = "guest-user",
+            linkedWorkspaceId = localWorkspaceId,
+            linkedEmail = null,
+            activeWorkspaceId = localWorkspaceId
+        )
+
+        try {
+            syncRepository.syncNow()
+        } catch (_: IllegalStateException) {
+        }
+
+        val localWorkspace = requireNotNull(environment.database.workspaceDao().loadAnyWorkspace()) {
+            "Expected a local workspace after guest normalization."
+        }
+        assertEquals(CloudAccountState.DISCONNECTED, environment.cloudPreferencesStore.currentCloudSettings().cloudState)
+        assertEquals(localWorkspaceId, environment.cloudPreferencesStore.currentCloudSettings().activeWorkspaceId)
+        assertEquals(localWorkspaceId, localWorkspace.workspaceId)
+        assertEquals(initialInstallationId, environment.cloudPreferencesStore.currentCloudSettings().installationId)
+        assertTrue(syncRepository.observeSyncStatus().first().status is SyncStatus.Failed)
+    }
+
+    @Test
+    fun syncNormalizesStaleActiveWorkspaceBeforeRunningWithoutReset() = runBlocking {
+        val syncRepository = environment.createSyncRepository(
+            remoteGateway = FakeCloudRemoteGateway.standard()
+        )
+        val localWorkspaceId = environment.requireLocalWorkspaceId()
+        val initialInstallationId = environment.cloudPreferencesStore.currentCloudSettings().installationId
+        environment.cloudPreferencesStore.updateCloudSettings(
+            cloudState = CloudAccountState.DISCONNECTED,
+            linkedUserId = null,
+            linkedWorkspaceId = null,
+            linkedEmail = null,
+            activeWorkspaceId = "stale-workspace-id"
+        )
+
+        try {
+            syncRepository.syncNow()
+        } catch (_: IllegalStateException) {
+        }
+
+        val cloudSettings = environment.cloudPreferencesStore.currentCloudSettings()
+        val localWorkspace = requireNotNull(environment.database.workspaceDao().loadAnyWorkspace()) {
+            "Expected a local workspace after stale active workspace normalization."
+        }
+        assertEquals(CloudAccountState.DISCONNECTED, cloudSettings.cloudState)
+        assertEquals(localWorkspaceId, cloudSettings.activeWorkspaceId)
+        assertEquals(localWorkspaceId, localWorkspace.workspaceId)
+        assertEquals(initialInstallationId, cloudSettings.installationId)
+        assertTrue(syncRepository.observeSyncStatus().first().status is SyncStatus.Failed)
+    }
+}
